+++ locales/es-ar/blocks.json
@@ -88,7 +88,8 @@
     "comment": "Describe el procedimiento...",
     "noParams": "",
     "addParam": "Agregar parámetro",
-    "addParamPrompt": "Ingresa el nombre del parámetro"
+    "addParamPrompt": "Ingresa el nombre del parámetro",
+    "variableName": "Nombre de variable"
   },
   "contextMenu": {
     "createProcedure": "Usar procedimiento '%1'",
